requirements.txt
# Web Framework
flask==3.0.0

# Environment Management
python-dotenv==1.0.0

# AI/ML
openai==1.12.0

# HTTP Requests
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4

# YAML Processing
pyyaml==6.0.1

# Excel Generation
openpyxl==3.1.2

# MCP (Model Context Protocol)
mcp==0.9.0

# Type Hints (for better code quality)
typing-extensions==4.9.0
